#40 Rename InWallSwitchHandler to LightControlHandler
This is the official name in the Bosch API documentation, so we should use it as well (https://apidocs.bosch-smarthome.com/local)
